Get correct StorageId while inserting 'Default folders' entries into db
When a guest/new user is created, while inserting entries of default
set of directories like Android, DCIM, Downloads etc of internal storage
into database, StorageId of internal storage specific to current user is
wrongly calculated, resulting in displaying nothing in MTP mode.
The cause for this is, the internal storage specific to user is always
created with storageId = 65537, but while querying, it is using the logic
((index + 1) << 16) + 1, and when we support multiple storages, the index
of internal storage increases, resulting in returning wrong StorageId.

Change-Id: I3fa39759b7a13cef7c5e248fb753647872e232c7